feat: add ability to use CDN or local scripts
Configuration is done by passing their base location to `LOCAL_SCRIPT_URL`
variable. If it set, then all of the scripts would come from local
instead, including the ones you've set to use the CDN ones

For local script, the name of the library is used as the filename for the
file containing it. With this we could create symlink it to the desired
version of that library