revert: 09ab316576a47d12578479b5f2e4e6fc3cfbe5dd
So the scenario was like this. The app wants to connect as `u` user to
use specific `sc` schema on `db` database hosted in `h`. The schema was
created by the root-like user.

On local setup (app -> database), this works fine simply by passing
postgres URI to the driver along with some runtime params via
query params (`search_path` for this case). However, if `h` is
remote (DBaaS, cloud, etc.), it might not work as expected.

Upon hours of researching, the reason mentioned in
09ab316576a47d12578479b5f2e4e6fc3cfbe5dd wasn't really the issue. The
actual factors found were:
- The query params were "ignored" when connecting
One of the scenario where this happens is when the DB was placed behind a
"connection pool manager". For PostgreSQL, PgBouncer is one of them and
apparently it doesn't support `search_path` parameter being passed (see:
https://github.com/pgbouncer/pgbouncer/issues/89).

- `search_path` wasn't set correctly
There are many reasons this could happen, such as the previous point. Be
sure user the `search_path` of `u` includes `sc` in it so that every
query would prioritize references to everything in it.

Special note in this project:
Actually, one could simply do a query to set `search_path` manually by
executing a query. However, `SET search_path` query is session-scoped,
meaning it's only available on one connection. If the connection was
lost and the app reconnects, the effect of that query would be gone.
Although untested, perhaps `ALTER USER` could be used to make the change
permanent. Going with this strategy, however, would make the app joins the
management of privileges in the DB. The more cooks, the worse the soup...

- `u` wasn't given enough privileges to `sc`
This will cause `u` to not being able to access `sc` and its tables. So,
one should remember to grant `u` privileges to them. For details, refer
to https://www.postgresql.org/docs/current/ddl-schemas.html

Among all of these causes, it had been decided that it's better to
leave the problem to be solved by the DBA guy. This scenario is not to
be handled by the application logic since it's more like a system
configuration problem